+++ clients/admin-ui/src/features/data-discovery-and-detection/action-center/tables/cells/DiscoveredAssetActionsCell.tsx
@@ -12,6 +12,7 @@ import { StagedResourceAPIResponse } from "~/types/api/models/StagedResourceAPIR
 import {
   useAddMonitorResultAssetsMutation,
   useIgnoreMonitorResultAssetsMutation,
+  useRestoreMonitorResultAssetsMutation,
 } from "../../action-center.slice";
 
 interface DiscoveredAssetActionsCellProps {
@@ -25,10 +26,15 @@ export const DiscoveredAssetActionsCell = ({
     useAddMonitorResultAssetsMutation();
   const [ignoreMonitorResultAssetsMutation, { isLoading: isIgnoringResults }] =
     useIgnoreMonitorResultAssetsMutation();
+  const [
+    restoreMonitorResultAssetsMutation,
+    { isLoading: isRestoringResults },
+  ] = useRestoreMonitorResultAssetsMutation();
 
   const { successAlert, errorAlert } = useAlert();
 
-  const anyActionIsLoading = isAddingResults || isIgnoringResults;
+  const anyActionIsLoading =
+    isAddingResults || isIgnoringResults || isRestoringResults;
 
   // eslint-disable-next-line @typescript-eslint/naming-convention
   const { urn, name, resource_type: type, diff_status } = asset;
@@ -61,35 +67,62 @@ export const DiscoveredAssetActionsCell = ({
     }
   };
 
+  const handleRestore = async () => {
+    const result = await restoreMonitorResultAssetsMutation({
+      urnList: [urn],
+    });
+    if (isErrorResult(result)) {
+      errorAlert(getErrorMessage(result.error));
+    } else {
+      successAlert(
+        `${type} "${name}" is no longer ignored and will appear in future scans.`,
+        `Restored`,
+      );
+    }
+  };
+
   // TODO [HJ-369] update disabled and tooltip logic once the categories of consent feature is implemented
   return (
     <Space>
-      <Tooltip
-        title={
-          !asset.system
-            ? `This asset requires a system before you can add it to the inventory.`
-            : undefined
-        }
-      >
-        <Button
-          data-testid="add-btn"
-          size="small"
-          onClick={handleAdd}
-          disabled={!asset.system || anyActionIsLoading}
-          loading={isAddingResults}
-        >
-          Add
-        </Button>
-      </Tooltip>
       {diff_status !== DiffStatus.MUTED && (
+        <>
+          <Tooltip
+            title={
+              !asset.system
+                ? `This asset requires a system before you can add it to the inventory.`
+                : undefined
+            }
+          >
+            <Button
+              data-testid="add-btn"
+              size="small"
+              onClick={handleAdd}
+              disabled={!asset.system || anyActionIsLoading}
+              loading={isAddingResults}
+            >
+              Add
+            </Button>
+          </Tooltip>
+          <Button
+            data-testid="ignore-btn"
+            size="small"
+            onClick={handleIgnore}
+            disabled={anyActionIsLoading}
+            loading={isIgnoringResults}
+          >
+            Ignore
+          </Button>
+        </>
+      )}
+      {diff_status === DiffStatus.MUTED && (
         <Button
-          data-testid="ignore-btn"
+          data-testid="restore-btn"
           size="small"
-          onClick={handleIgnore}
+          onClick={handleRestore}
           disabled={anyActionIsLoading}
-          loading={isIgnoringResults}
+          loading={isRestoringResults}
         >
-          Ignore
+          Restore
         </Button>
       )}
     </Space>