+++ tests/ctl/core/test_api_helpers.py
@@ -20,7 +20,8 @@ PARAM_MODEL_LIST = [
 
 
 # Fixtures
-@pytest.fixture
+@pytest.fixture(scope="function")
+@pytest.mark.usefixtures("monkeypatch_requests")
 def created_resources(
     test_config: FidesConfig, resources_dict: Dict, request: FixtureRequest
 ) -> Generator:
@@ -77,10 +78,13 @@ def delete_resource_type(test_config: FidesConfig, resource_type: str) -> None:
 
 
 @pytest.mark.integration
+@pytest.mark.usefixtures("monkeypatch_requests")
 class TestGetServerResource:
+
     @pytest.mark.parametrize(
         "created_resources", PARAM_MODEL_LIST, indirect=["created_resources"]
     )
+    @pytest.mark.usefixtures("fideslang_resources")
     def test_get_server_resource_found_resource(
         self, test_config: FidesConfig, created_resources: List
     ) -> None:
@@ -116,10 +120,10 @@ class TestGetServerResource:
 
 @pytest.mark.integration
 class TestGetServerResources:
-    @pytest.mark.integration
     @pytest.mark.parametrize(
         "created_resources", PARAM_MODEL_LIST, indirect=["created_resources"]
     )
+    @pytest.mark.usefixtures("monkeypatch_requests", "fideslang_resources")
     def test_get_server_resources_found_resources(
         self, test_config: FidesConfig, created_resources: List
     ) -> None:
@@ -156,6 +160,7 @@ class TestGetServerResources:
 
 @pytest.mark.integration
 class TestListServerResources:
+    @pytest.mark.usefixtures("monkeypatch_requests", "fideslang_resources")
     def test_list_server_resources_passing(self, test_config: FidesConfig) -> None:
         resource_type = "data_category"
         result = _api_helpers.list_server_resources(