+++ CHANGELOG.md
@@ -21,10 +21,11 @@ Changes can also be flagged with a GitHub label for tracking purposes. The URL o
 
 ## [Unreleased](https://github.com/ethyca/fides/compare/2.60.1...main)
 
+### Added
 - Added Recorded URL to Consent Report [#6077](https://github.com/ethyca/fides/pull/6077)
 
-
-
+### Developer Experience
+- Cleaning up test fixtures [#6008](https://github.com/ethyca/fides/pull/6008)
 
 ## [2.60.0](https://github.com/ethyca/fides/compare/2.59.2...2.60.0)